requirements.txt
Package                Version
---------------------- -------
click                  8.1.2
colorama               0.4.4
dnspython              2.2.1
email-validator        1.1.3
Flask                  2.0.0
Flask-SQLAlchemy       2.5.1
Flask-WTF              1.0.1
greenlet               1.1.2
idna                   3.3
itsdangerous           2.1.2
Jinja2                 3.1.1
MarkupSafe             2.1.1
mysql-connector-python 8.0.28
pip                    21.1.2
protobuf               3.20.0
setuptools             57.0.0
SQLAlchemy             1.4.35
Werkzeug               2.1.1
wheel                  0.36.2
WTForms                3.0.1
